third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/navigation_unload_same_origin-expected.txt

This is a testharness.js-based test.
FAIL Same-origin navigation started from unload handler assert_equals: expected "001-3" but got "001-2"
Harness: the test ran to completion.
